refactor(creature): remove MonsterListScreen and rename Compact variants
